Fixed movements (had to remove all interpolations and predictions from simulation)
git-svn-id: https://openlierox.svn.sourceforge.net/svnroot/openlierox@1211 34602234-ff1f-0410-a465-ea8f3b77ab7f

+++ src/client/CClient_Game.cpp
@@ -136,9 +136,9 @@ void CClient::Simulation(void)
 			w->Simulate(cRemoteWorms, local, tLX->fDeltaTime);
 			
 			// Get the position the worm is likely to be in on the server
-			// TODO: does not work
+			// TODO: does not work as expected
 			/*CVec oldpos = w->getPos();
-			w->Simulate(cRemoteWorms, local, iMyPing / 1000); // I shoot myself regularly, so added some coef here 
+			w->Simulate(cRemoteWorms, local, (float)iMyPing / 1000);
 			w->setNextPos(w->getPos());
 			w->setPos(oldpos);*/
